REFACTOR: Custom Exception 리팩토링
