+++ src/main/java/kr/hhplus/be/server/application/payment/facade/PaymentFacade.java
@@ -8,8 +8,8 @@ import kr.hhplus.be.server.domain.order.service.OrderService;
 import kr.hhplus.be.server.domain.payment.dto.PaymentResult;
 import kr.hhplus.be.server.domain.payment.enums.PaymentStatusType;
 import kr.hhplus.be.server.domain.payment.service.PaymentService;
+import kr.hhplus.be.server.support.exception.CommonException;
 import kr.hhplus.be.server.support.exception.payment.PaymentErrorCode;
-import kr.hhplus.be.server.support.exception.payment.PaymentException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,7 +28,7 @@ public class PaymentFacade {
         OrderResult order = orderService.findOrderInfoNoProduct(orderId);
 
         if(OrderStateType.PAYED.equals(order.getOrderState())) {
-            throw new PaymentException(PaymentErrorCode.INVALID_ORDER_STATE);
+            throw new CommonException(PaymentErrorCode.INVALID_ORDER_STATE);
         }
 
         PaymentResult paymentResult = paymentService.createPayment(request.toServiceRequest(order.getSalePrice(), PaymentStatusType.PAYED));
