+++ src/main/java/kr/hhplus/be/server/domain/point/entity/UserPoint.java
@@ -5,8 +5,8 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import kr.hhplus.be.server.domain.BaseEntity;
+import kr.hhplus.be.server.support.exception.CommonException;
 import kr.hhplus.be.server.support.exception.point.PointErrorCode;
-import kr.hhplus.be.server.support.exception.point.PointException;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -38,7 +38,7 @@ public class UserPoint extends BaseEntity {
 
     public void decrease(Long amount){
         if(this.amount < amount){
-            throw new PointException(PointErrorCode.USE_POINT_LACK);
+            throw new CommonException(PointErrorCode.USE_POINT_LACK);
         }
         this.amount -= amount;
     }
