+++ src/main/java/kr/hhplus/be/server/domain/user/service/UserService.java
@@ -3,8 +3,8 @@ package kr.hhplus.be.server.domain.user.service;
 import kr.hhplus.be.server.domain.user.dto.UserResult;
 import kr.hhplus.be.server.domain.user.entity.User;
 import kr.hhplus.be.server.domain.user.repository.UserRepository;
+import kr.hhplus.be.server.support.exception.CommonException;
 import kr.hhplus.be.server.support.exception.user.UserErrorCode;
-import kr.hhplus.be.server.support.exception.user.UserException;
 import kr.hhplus.be.server.support.util.user.UserValidationUtils;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.ObjectUtils;
@@ -24,7 +24,7 @@ public class UserService {
         User user = userRepository.findByUserId(userId);
 
         if(ObjectUtils.isEmpty(user)) {
-            throw new UserException(UserErrorCode.INVALID_USER);
+            throw new CommonException(UserErrorCode.INVALID_USER);
         }
 
         return UserResult.toResult(user);
