+++ src/main/java/kr/hhplus/be/server/support/exception/CommonException.java
@@ -1,14 +1,14 @@
-package kr.hhplus.be.server.support.exception.order;
+package kr.hhplus.be.server.support.exception;
 
 import kr.hhplus.be.server.support.ErrorCode;
 import lombok.Getter;
 
 @Getter
-public class OrderException extends RuntimeException {
+public class CommonException extends IllegalStateException {
     private final ErrorCode errorCode;
     private final Object[] args;
 
-    public OrderException(ErrorCode errorCode, Object... args) {
+    public CommonException(ErrorCode errorCode, Object... args) {
         super(errorCode.getMessage());
         this.errorCode = errorCode;
         this.args = args;
