+++ src/main/java/kr/hhplus/be/server/support/util/user/UserValidationUtils.java
@@ -1,13 +1,13 @@
 package kr.hhplus.be.server.support.util.user;
 
+import kr.hhplus.be.server.support.exception.CommonException;
 import kr.hhplus.be.server.support.exception.user.UserErrorCode;
-import kr.hhplus.be.server.support.exception.user.UserException;
 
 public class UserValidationUtils {
 
     public static void validateUserId(Long id) {
         if (id == null || id <= 0) {
-            throw new UserException(UserErrorCode.USER_ID_LESS_THAN_ZERO);
+            throw new CommonException(UserErrorCode.USER_ID_LESS_THAN_ZERO);
         }
     }
 }
